requirements.txt
python-telegram-bot==21.6
requests==2.32.3
redis==5.0.8
pytz==2025.2
python-dotenv==1.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
celery==5.3.4
celery[redis]==5.3.4
slowapi==0.1.9
gunicorn==21.2.0
